mucho cleano

+++ apps/keyforge-hive/src/cron.rs
@@ -20,12 +20,12 @@ use tracing::{error, info};
 
 // --- Defaults ---
 pub const DEFAULT_CRON_INTERVAL_SECS: u64 = 60;
-pub const DEFAULT_ZOMBIE_TIMEOUT_MINS: i64 = 70;
+pub const DEFAULT_ZOMBIE_TIMEOUT_MINS: i32 = 70;
 pub const DEFAULT_ZOMBIE_MAX_RETRIES: i32 = 3;
-pub const DEFAULT_PRUNE_OLD_JOBS_DAYS: i64 = 30;
-pub const DEFAULT_PRUNE_INACTIVE_NODES_MINS: i64 = 15;
-pub const DEFAULT_PRUNE_RESULTS_DAYS: i64 = 7;
-pub const DEFAULT_PRUNE_RESULTS_KEEP_COUNT: i64 = 1000;
+pub const DEFAULT_PRUNE_OLD_JOBS_DAYS: i32 = 30;
+pub const DEFAULT_PRUNE_INACTIVE_NODES_MINS: i32 = 15;
+pub const DEFAULT_PRUNE_RESULTS_DAYS: i32 = 7;
+pub const DEFAULT_PRUNE_RESULTS_KEEP_COUNT: i32 = 1000;
 
 /// Spawns background maintenance tasks that run periodically.
 ///
@@ -47,7 +47,7 @@ pub async fn start_cron_jobs(
         hour_ticker += 1;
 
         // --- 1. THE REAPER (Runs every minute) ---
-        match job_repo.prune_stale_jobs(DEFAULT_ZOMBIE_TIMEOUT_MINS as i32, DEFAULT_ZOMBIE_MAX_RETRIES).await {
+        match job_repo.prune_stale_jobs(DEFAULT_ZOMBIE_TIMEOUT_MINS, DEFAULT_ZOMBIE_MAX_RETRIES).await {
             Ok(count) if count > 0 => {
                 tracing::warn!("💀 Zombie Reaper: Reset {} stuck jobs.", count);
             }
@@ -61,17 +61,17 @@ pub async fn start_cron_jobs(
             info!("🧹 Running hourly maintenance...");
 
             // Cleanup old jobs (cancelled/stale > 30 days)
-            if let Err(e) = job_repo.prune_old_jobs(DEFAULT_PRUNE_OLD_JOBS_DAYS as i32).await {
+            if let Err(e) = job_repo.prune_old_jobs(DEFAULT_PRUNE_OLD_JOBS_DAYS).await {
                 error!("Failed to prune old jobs: {}", e);
             }
 
             // Cleanup inactive nodes (> 15 mins)
-            if let Err(e) = node_repo.prune_inactive_nodes(DEFAULT_PRUNE_INACTIVE_NODES_MINS as i32).await {
+            if let Err(e) = node_repo.prune_inactive_nodes(DEFAULT_PRUNE_INACTIVE_NODES_MINS).await {
                 error!("Failed to prune inactive nodes: {}", e);
             }
 
             // Prune Results (Keep top 1000 per job, delete others older than 7 days)
-            match result_repo.prune_old_results(DEFAULT_PRUNE_RESULTS_DAYS as i32, DEFAULT_PRUNE_RESULTS_KEEP_COUNT as i32).await {
+            match result_repo.prune_old_results(DEFAULT_PRUNE_RESULTS_DAYS, DEFAULT_PRUNE_RESULTS_KEEP_COUNT).await {
                 Ok(count) => {
                     if count > 0 {
                         info!("🧹 Pruned {} old results.", count);
